fix ee ui
